chore: move jsx tokens into configs dir
