+++ packages/sage-react/lib/configs/classnames/grid-templates.js
@@ -1,4 +1,4 @@
-import { SageDictionary } from '../../dictionary';
+import { SageDictionary } from '../dictionary';
 
 // CLASSNAMES_GRID_TEMPLATES.ETE --> `sage-grid-template-ete`
 export const CLASSNAMES_GRID_TEMPLATES = {};
