+++ packages/sage-react/lib/configs/dictionary/index.js
